requirements.txt
discord.py
Flask
flask-cors
